prep for test

+++ opendbc/safety/tests/libsafety/safety_helpers.py
@@ -15,9 +15,9 @@ def setup_safety_helpers(ffi):
   bool get_brake_pressed_prev(void);
   bool get_regen_braking_prev(void);
   bool get_acc_main_on(void);
-  int get_vehicle_speed_min(void);
-  int get_vehicle_speed_max(void);
-  int get_vehicle_speed_last(void);
+  float get_vehicle_speed_min(void);
+  float get_vehicle_speed_max(void);
+  float get_vehicle_speed_last(void);
   int get_current_safety_mode(void);
   int get_current_safety_param(void);
